Rename pipeline to workflow. Move files and add utility module
