--- src/vivarium_cluster_tools/psimulate/pipeline_config/__init__.py
@@ -1,14 +1,0 @@
-"""
-========================
-Pipeline Configuration
-========================
-
-Configuration and utilities for pipeline workflow orchestration.
-
-"""
-
-from vivarium_cluster_tools.psimulate.pipeline_config.config import (
-    PipelineConfig,
-    ResourceConfig,
-    StepConfig,
-)
